requirements.txt
# requirements.txt
kafka-python
valkey
requests
python-dotenv
psycopg2-binary
aiohttp
duckdb
pandas
lnhistoryclient==3.6.0